fix drop users table migration

+++ db/migrate/20171015170144_drop_users.rb
@@ -1,5 +1,11 @@
 class DropUsers < ActiveRecord::Migration[5.0]
-  def change
-    drop_table :users
+  def up
+    if ActiveRecord::Base.connection.table_exists? :users
+      drop_table :users
+    end
+  end
+
+  def down
+
   end
 end
